refactor: Split ioapic and apic in multiple files but initialize with common init function
